Refactor pay element updating so that it actually saves
The ef core identity mapper tracks instances of entities but since we
weren't actually changing any entities it didn't save anything when
calling SaveChangesAsync. Even the new pay element didn't get saved
because we constructed the entity manually without adding it to a DbSet.

To fix this the code is refactored to more like the updating of pay
elements in the api application where it uses the PayElements property
of the timesheet as the container. This ensures the identity mapper
sees the changes when SaveChangesAsync is called. It also allows us to
remove the PayElementsGateway since everything is mediated through the
timesheet and the PayElementMapper doesn't need the timesheet since
the identity mapper takes care of assigning the foreign keys.